refactor(subscribers): извлечён SubscribersResource (#210 ось видимости, PR3)
Третий (последний) ресурс оси scope унифицирован. Извлечён `SubscribersResource({scope,scopeId})`
— тулбар «Сообщение» (Admin) + список подписчиков + добавление + SendMessageDialog, без чипа
области и внешнего коллапса. `SubscribersPanel` сведён к тонкой коллапс-обёртке над ним (коллапс+чип
уйдут при постройке scope-страниц). Кнопка «Сообщение» переехала из шапки коллапса в тулбар ресурса.
